Fix native handle data type from int to IntPtr
and care for uint parameters

+++ FastIOW/Boards/IOWarrior28.cs
@@ -1,23 +1,24 @@
 ﻿/*
- *   
+ *
  *   Copyright 2020 Florian Porsch <tederean@gmail.com>
- *   
+ *
  *   This program is free software; you can redistribute it and/or modify
  *   it under the terms of the GNU Lesser General Public License as published by
  *   the Free Software Foundation; either version 3 of the License, or
  *   (at your option) any later version.
- *   
+ *
  *   This program is distributed in the hope that it will be useful,
  *   but WITHOUT ANY WARRANTY; without even the implied warranty of
  *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *   GNU Lesser General Public License for more details.
- *   
+ *
  *   You should have received a copy of the GNU Lesser General Public License
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
  *   MA 02110-1301 USA.
  *
  */
+using System;
 using Tederean.FastIOW.Internal;
 
 namespace Tederean.FastIOW
@@ -100,7 +101,7 @@ namespace Tederean.FastIOW
     public const int BUILDIN_BUTTON = P2_1;
 
 
-    internal IOWarrior28(int handle) : base(handle)
+    internal IOWarrior28(IntPtr handle) : base(handle)
     {
       I2C = new I2CInterfaceImplementation(this, Pipe.I2C_MODE, 62);
       ADC = new ADCInterfaceImplementation(this, Pipe.ADC_MODE, AnalogPins);
